\name{bulls}
\docType{data}
\alias{bulls}
\title{Bulls Characteristics}
\description{
The amount of daughters' milk, the amount of fat in this milk, height and body length. All values are differences to the average.
}
\source{Institute of Animal Science in Prague, Plemdat s.r.o.}
\keyword{datasets}
