\name{models.and.types}
\alias{models.and.types}
\alias{MODEL} 
\alias{TYPE}
\alias{TYPE.OF.MODEL}
\alias{npar} 
\alias{GRAPHICAL}
\alias{GR.TR}
\title{Models and Types}
\description{Basic data structures representing models and types. 
}
\details{
\code{MODEL} and \code{TYPE} are vectors of independence characters; \code{TYPE.OF.MODEL} is a model-to-type transfering table; \code{npar} reflects an effective dimension of the type; \code{GRAPHICAL} is TRUE/FALSE vector saying which model is graphical; \code{GR.TR} is the same thing but regarding types.
}
\keyword{internal}
