\name{is.positive.definite}
\alias{is.positive.definite}
\alias{is.positive.definite} 
\title{Test for Positive Definiteness}
\description{Test whether a matrix is positive definite by computing its eigen values.
}
\details{
This function was taken from a package \code{corpcor}.
}
\keyword{internal}
