\name{type.permutations}
\alias{type.permutations}
\alias{PERM} 
\alias{INVPERM} 
\alias{TR.MODEL} 
\alias{TR.TYPE} 
\title{Isomorphisms}
\description{
How to obtain model as a permutation of some type.
}
\details{
These structures are taking care about model-to-type translation after all possible permtations of variables (\code{PERM}). \code{INVPERM} are inverse permutations to \code{PERM}.
}
\keyword{internal}
