\name{types.fitting}
\alias{types.fitting}
\alias{mle} 
\alias{mle1} 
\alias{mle2} 
\alias{mle3} 
\alias{mle4} 
\alias{mle5} 
\alias{mle6} 
\alias{mle7} 
\alias{mle8} 
\alias{mle9} 
\alias{mle10} 
\alias{mle11}
\alias{mle12} 
\alias{mle13} 
\alias{mle14} 
\alias{mle15} 
\alias{mle16} 
\alias{mle17} 
\alias{mle18} 
\alias{mle19} 
\alias{mle20} 
\alias{mle21}
\alias{mle22} 
\alias{mle23} 
\alias{mle24} 
\alias{mle25} 
\alias{mle26} 
\alias{mle27} 
\alias{mle28} 
\alias{mle29} 
\alias{mle30} 
\alias{mle31}
\alias{mle32} 
\alias{mle33} 
\alias{mle34} 
\alias{mle35} 
\alias{mle36} 
\alias{mle37} 
\alias{mle38} 
\alias{mle39} 
\alias{mle40} 
\alias{mle41}
\alias{mle42} 
\alias{mle43} 
\alias{mle44} 
\alias{mle45} 
\alias{mle46} 
\alias{mle47} 
\alias{mle48} 
\alias{mle49} 
\alias{mle50} 
\alias{mle51}
\alias{mle52} 
\alias{mle53} 
\title{Fitting 53 Representable Types}
\description{The list \code{mle} consists of 53 functions \code{mle1}-\code{mle53} fitting 53 types that are representable by a regular Gaussian distribution.
}
\details{
Anybody is welcome to improve these functions.
}
\seealso{\code{\link{types.sampling}}, \code{\link{ind.mle}}}
\keyword{internal}
