\name{types.sampling}
\alias{types.sampling}
\alias{gen} 
\alias{gen1} 
\alias{gen2} 
\alias{gen3} 
\alias{gen4} 
\alias{gen5} 
\alias{gen6} 
\alias{gen7} 
\alias{gen8} 
\alias{gen9} 
\alias{gen10} 
\alias{gen11}
\alias{gen12} 
\alias{gen13} 
\alias{gen14} 
\alias{gen15} 
\alias{gen16} 
\alias{gen17} 
\alias{gen18}
\alias{gen19} 
\alias{gen20} 
\alias{gen21} 
\alias{gen22} 
\alias{gen23} 
\alias{gen24} 
\alias{gen25} 
\alias{gen26} 
\alias{gen27} 
\alias{gen28}
\alias{gen29} 
\alias{gen30} 
\alias{gen31} 
\alias{gen32} 
\alias{gen33} 
\alias{gen34} 
\alias{gen35} 
\alias{gen36} 
\alias{gen37} 
\alias{gen38}
\alias{gen39} 
\alias{gen40} 
\alias{gen41} 
\alias{gen42} 
\alias{gen43} 
\alias{gen44} 
\alias{gen45} 
\alias{gen46} 
\alias{gen47} 
\alias{gen48}
\alias{gen49} 
\alias{gen50} 
\alias{gen51} 
\alias{gen52} 
\alias{gen53} 
\title{Generating Random Variance Matrix for 53 Representable Types}
\description{The list \code{gen} consists of 53 functions \code{gen1}-\code{gen53} generating random matrices corresponding to 53 types that are representable by a regular Gaussian distribution.
}
\seealso{\code{\link{types.fitting}}, \code{\link{ind.rgauss}}}
\keyword{internal}

