\name{yet.undocumented}
\alias{yet.undocumented}
\alias{CI} 
\alias{gen} 
\alias{gen1} 
\alias{gen10} 
\alias{gen11}
\alias{gen12} 
\alias{gen13} 
\alias{gen14} 
\alias{gen15} 
\alias{gen16} 
\alias{gen17} 
\alias{gen18}
\alias{gen19} 
\alias{gen2} 
\alias{gen20} 
\alias{gen21} 
\alias{gen22} 
\alias{gen23} 
\alias{gen24} 
\alias{gen25} 
\alias{gen26} 
\alias{gen27} 
\alias{gen28}
\alias{gen29} 
\alias{gen3} 
\alias{gen30} 
\alias{gen31} 
\alias{gen32} 
\alias{gen33} 
\alias{gen34} 
\alias{gen35} 
\alias{gen36} 
\alias{gen37} 
\alias{gen38}
\alias{gen39} 
\alias{gen4} 
\alias{gen40} 
\alias{gen41} 
\alias{gen42} 
\alias{gen43} 
\alias{gen44} 
\alias{gen45} 
\alias{gen46} 
\alias{gen47} 
\alias{gen48}
\alias{gen49} 
\alias{gen5} 
\alias{gen50} 
\alias{gen51} 
\alias{gen52} 
\alias{gen53} 
\alias{gen6} 
\alias{gen7} 
\alias{gen8} 
\alias{gen9} 
\alias{GRAPHICAL}
\alias{GR.TR} 
\alias{ind.test} 
\alias{INVPERM} 
\alias{is.positive.definite} 
\alias{mle} 
\alias{mle1} 
\alias{mle10} 
\alias{mle11}
\alias{mle12} 
\alias{mle13} 
\alias{mle14} 
\alias{mle15} 
\alias{mle16} 
\alias{mle17} 
\alias{mle18} 
\alias{mle19} 
\alias{mle2} 
\alias{mle20} 
\alias{mle21}
\alias{mle22} 
\alias{mle23} 
\alias{mle24} 
\alias{mle25} 
\alias{mle26} 
\alias{mle27} 
\alias{mle28} 
\alias{mle29} 
\alias{mle3} 
\alias{mle30} 
\alias{mle31}
\alias{mle32} 
\alias{mle33} 
\alias{mle34} 
\alias{mle35} 
\alias{mle36} 
\alias{mle37} 
\alias{mle38} 
\alias{mle39} 
\alias{mle4} 
\alias{mle40} 
\alias{mle41}
\alias{mle42} 
\alias{mle43} 
\alias{mle44} 
\alias{mle45} 
\alias{mle46} 
\alias{mle47} 
\alias{mle48} 
\alias{mle49} 
\alias{mle5} 
\alias{mle50} 
\alias{mle51}
\alias{mle52} 
\alias{mle53} 
\alias{mle6} 
\alias{mle7} 
\alias{mle8} 
\alias{mle9} 
\alias{MODEL} 
\alias{npar} 
\alias{PERM} 
\alias{TR.MODEL} 
\alias{TR.TYPE} 
\alias{TYPE}
\alias{TYPE.OF.MODEL}
\title{Internal Functions}
\description{To be documented later.
}
\details{
To be added later.
}
\keyword{internal}
